normalize fields prior to injection
I was trying to use this latest update for custom endpoints to create Gridsome Collections based on a specific WP endpoint. I was trying to query `route: "/wp/v2/posts?categories=251"` but it was giving me an error. From the source code it seems like you ran into the same one, hence the "normalizeFields" function. I implemented it into the getCustomEndpoints function and it works great!