Added option for multiple bases
When using Airtable as a CMS, one might want to modularize its content in separate bases for organization or security purposes, or even to try not to exceed Airtable's record-per-base limit. This modification makes it possible to connect multiple bases to Gridsome, each one with its tables and other attributes we already know and use.
I didn't run tests or anything, but I'm succesfully using this adjustment in production for over 3 months now.